+++ firebase-firestore/src/main/java/com/google/firebase/firestore/local/LocalStore.java
@@ -128,7 +128,8 @@ public final class LocalStore {
     targetIdGenerator = TargetIdGenerator.forQueryCache(queryCache.getHighestTargetId());
     mutationQueue = persistence.getMutationQueue(initialUser);
     remoteDocuments = persistence.getRemoteDocumentCache();
-    localDocuments = new LocalDocumentsView(remoteDocuments, mutationQueue);
+    localDocuments =
+        new LocalDocumentsView(remoteDocuments, mutationQueue, persistence.getIndexManager());
     // TODO: Use IndexedQueryEngine as appropriate.
     queryEngine = new SimpleQueryEngine(localDocuments);
 
@@ -162,7 +163,8 @@ public final class LocalStore {
     List<MutationBatch> newBatches = mutationQueue.getAllMutationBatches();
 
     // Recreate our LocalDocumentsView using the new MutationQueue.
-    localDocuments = new LocalDocumentsView(remoteDocuments, mutationQueue);
+    localDocuments =
+        new LocalDocumentsView(remoteDocuments, mutationQueue, persistence.getIndexManager());
     // TODO: Use IndexedQueryEngine as appropriate.
     queryEngine = new SimpleQueryEngine(localDocuments);