+++ firebase-firestore/src/main/java/com/google/firebase/firestore/local/SQLitePersistence.java
@@ -80,6 +80,7 @@ public final class SQLitePersistence extends Persistence {
   private File databasePath;
   private boolean started;
   private final SQLiteQueryCache queryCache;
+  private final SQLiteIndexManager indexManager;
   private final SQLiteRemoteDocumentCache remoteDocumentCache;
   private final SQLiteLruReferenceDelegate referenceDelegate;
   private final SQLiteTransactionListener transactionListener =
@@ -109,6 +110,7 @@ public final class SQLitePersistence extends Persistence {
     this.databasePath = context.getDatabasePath(databaseName);
     this.serializer = serializer;
     this.queryCache = new SQLiteQueryCache(this, this.serializer);
+    this.indexManager = new SQLiteIndexManager(this);
     this.remoteDocumentCache = new SQLiteRemoteDocumentCache(this, this.serializer);
     this.referenceDelegate = new SQLiteLruReferenceDelegate(this, params);
   }
@@ -161,6 +163,11 @@ public final class SQLitePersistence extends Persistence {
   @Override
   SQLiteQueryCache getQueryCache() {
     return queryCache;
+  }
+
+  @Override
+  IndexManager getIndexManager() {
+    return indexManager;
   }
 
   @Override